chore(data-workflows): Created CollectionListing for GbqToGcs
