+++ src/components/data-workflows/runs-collections-listings.ts
@@ -1,0 +1,3 @@
+import GbqToGcs from './gbq-to-gcs/runs/CollectionListing.vue';
+
+export const GbqToGcsCollectionListing = GbqToGcs;
