+++ src/views/data-workflows/gbq-to-gcs/runs/ListingView.vue
@@ -1,25 +1,7 @@
 <template>
 	<div>
 		<data-management-header :workflowName="workflowName" :tabsItems="tabsItems" />
-		<listing-component
-			:type="listingType"
-			:module-name="moduleName"
-			:headers="headers"
-			:overridden-columns="overriddenColumns"
-			sort-by="dag_execution_date"
-			:sort-desc="true"
-			show-airflow-action
-		>
-			<template v-slot:gcs_triggering_file="{ item: { id, gcs_triggering_file } }">
-				<router-link :to="{ name: routeName, params: { id } }">
-					<span class="font-weight-medium">{{ gcs_triggering_file }}</span>
-				</router-link>
-			</template>
-
-			<template v-slot:dag_execution_date="{ item: { dag_execution_date } }">
-				{{ dag_execution_date | moment('YYYY/MM/DD - HH:mm') }}
-			</template>
-		</listing-component>
+		<collection-listing />
 	</div>
 </template>
 
@@ -27,31 +9,11 @@
 import { Component, Mixins } from 'vue-property-decorator';
 
 import HeaderInfosMixin from '../header-infos';
-import RunCollectionMixin from '@/mixins/data-workflows/collection/run-collection-mixin';
+import CollectionMixin from '@/mixins/data-workflows/collection/collection-mixin';
+import CollectionListing from '@/components/data-workflows/gbq-to-gcs/runs/CollectionListing.vue';
 
-import { getGbqToGcsRuns } from '@/store/modules/easy-firestore/get-gbq-to-gcs-runs';
-import { GBQ_TO_GCS_RUNS_ITEM } from '@/constants/router/routes-names';
-import {
-	ACCOUNT,
-	ENVIRONMENT,
-	DESTINATION_BUCKET,
-	GCS_TRIGGERING_FILE,
-	STATUS,
-	DAG_EXECUTION_DATE,
-	ACTIONS
-} from '@/constants/data-workflows/listing/header-items';
-
-@Component
-export default class GbqToGcsRunsListingView extends Mixins(HeaderInfosMixin, RunCollectionMixin) {
-	moduleName: string = getGbqToGcsRuns.moduleName;
-	overriddenColumns: string[] = ['gcs_triggering_file', 'dag_execution_date'];
-
-	get routeName() {
-		return GBQ_TO_GCS_RUNS_ITEM;
-	}
-
-	get headers() {
-		return [ACCOUNT, ENVIRONMENT, DESTINATION_BUCKET, GCS_TRIGGERING_FILE, STATUS, DAG_EXECUTION_DATE, ACTIONS];
-	}
-}
+@Component({
+	components: { CollectionListing }
+})
+export default class GbqToGcsRunsListingView extends Mixins(HeaderInfosMixin, CollectionMixin) {}
 </script>
